Apply(feature) : create topic with images

+++ app/Views/admin/board/topic_input.php
@@ -1,10 +1,19 @@
 <?php
+
+use Crisu83\ShortId\ShortId;
+
+if ($type == 'create') {
+    $title = '';
+    $content = '';
+}
+$shortid = ShortId::create();
+$identifier = $shortid->generate();
 ?>
 
 <div class="container-inner">
     <div class="inner-box">
         <h3 class="title">
-            Notice
+            <?= $alias ?>
         </h3>
         <div class="topic-wrap">
             <div class="form-box">
@@ -15,6 +24,8 @@
                 <div class="text-wrap line-after">
                     <textarea placeholder="Content" name="content" class="content editable"><?= $content ?></textarea>
                 </div>
+                <input hidden type="text" name="board_id" class="editable" value="<?= $board_id ?>"/>
+                <input hidden type="text" name="identifier" class="editable" value="<?= $identifier ?>"/>
             </div>
             <div class="slider-wrap">
                 <div class="slider-box">
@@ -22,37 +33,41 @@
                         <div class="slick-element add"
                              style="background: url('/asset/images/icon/plus_circle_big.png') no-repeat center; font-size: 0;">
                             <label for="file" class="button"></label>
-                            <input type="file" name="file" multiple id="file" onchange="onFileUpload(this);"
+                            <input type="file" name="file" multiple id="file"
+                                   onchange="onFileUpload(this,'<?= $identifier ?>');"
                                    accept="image/*"/>
                             <!--                            <a href="#" class="button"></a>-->
                         </div>
-                        <?php foreach ($images as $index => $image) { ?>
-                            <div class="slick-element"
-                                 style="background: url('<?= $image ?>') no-repeat center; background-size: cover; font-size: 0;">
-                                Slider #<?= $index ?>
-                                <div class="slick-element-hover">
-                                    <a href="javascript:deleteImage(<?= $index + 1 ?>)"
-                                       class="button delete-image black">
-                                        <img src="/asset/images/icon/cancel_white.png"/>
-                                    </a>
+                        <?php if (isset($images)) {
+                            foreach ($images as $index => $image) { ?>
+                                <script>
+                                    image_file_ids.push('<?=$image['id']?>');
+                                </script>
+                                <div class="slick-element"
+                                     style="background: url('<?= $image ?>') no-repeat center; background-size: cover; font-size: 0;">
+                                    Slider #<?= $index ?>
+                                    <div class="slick-element-hover">
+                                        <a href="javascript:deleteImage(<?= $index + 1 ?>)"
+                                           class="button delete-image black">
+                                            <img src="/asset/images/icon/cancel_white.png"/>
+                                        </a>
+                                    </div>
                                 </div>
-                            </div>
-                        <?php } ?>
-
-                        <div class="slick-element"
-                             style="background: url('./writable/uploads/object.png') no-repeat center; background-size: cover; font-size: 0;">
-                            Slider #1
-                        </div>
+                            <?php }
+                        } ?>
                     </div>
                 </div>
             </div>
             <div class="button-wrap">
-                <a href="javascript:confirmEditTopic(<?= $id ?>)" class="button confirm black">Confirm</a>
+                <a href="<?= $type == 'create' ? 'javascript:confirmCreateTopic()' : 'javascript:confirmEditTopic(' . $id . ')' ?>"
+                   class="button confirm black">Confirm</a>
             </div>
         </div>
     </div>
 </div>
 <script type="text/javascript">
+    let image_file_ids = [];
+
     $('.slider-wrap .slick').slick({
         slidesToShow: 4,
         slidesToScroll: 1,
@@ -69,7 +84,10 @@
             dataType: 'json',
             url: `/api/topic/update/${id}`,
             success: function (response, status, request) {
-                //TODO refresh
+                if (!response.success) {
+                    openPopupErrors('popup-error', response, status, request);
+                    return;
+                }
             },
             error: function (response, status, error) {
                 console.log(error)
@@ -77,11 +95,50 @@
         });
     }
 
-    function deleteImage(index) {
-        $('.slider-wrap .slick').slick('slickRemove', index);
+    function confirmCreateTopic() {
+        let data = parseInputToData($(`.form-box .editable`))
+
+        data['images'] = image_file_ids;
+
+        $.ajax({
+            type: 'POST',
+            url: `/api/topic/create`,
+            data: data,
+            dataType: 'json',
+            success: function (response, status, request) {
+                if (!response.success) {
+                    openPopupErrors('popup-error', response, status, request);
+                    return;
+                }
+                History.back();
+            },
+            error: function (response, status, error) {
+                openPopupErrors('popup-error', response, status, error);
+            },
+        });
     }
 
-    function onFileUpload(input, id) {
+    function deleteImage(id) {
+        let index = image_file_ids.indexOf(id);
+        if (index >= 0) {
+            $('.slider-wrap .slick').slick('slickRemove', index + 1);
+            image_file_ids.splice(index, 1);
+            $.ajax({
+                type: 'DELETE',
+                url: `/api/image-file/delete/${id}`,
+                dataType: 'json',
+                success: function (response, status, request) {
+                    openPopupErrors('popup-error', response, status, request);
+                },
+                error: function (response, status, error) {
+                    openPopupErrors('popup-error', response, status, error);
+                },
+            });
+
+        }
+    }
+
+    function onFileUpload(input, identifier) {
         if (input.files.length == 0) return;
         let form = new FormData();
         for (let i in input.files) {
@@ -91,33 +148,49 @@
 
         $.ajax({
             type: 'POST',
-            url: `/api/image-file/upload`,
+            url: `/api/image-file/upload/${identifier}`,
             data: form,
             processData: false,
             contentType: false,
             cache: false,
-            dataType: "multipart/form-data",
+            dataType: "json",
             success: function (response, status, request) {
-                console.log(response)
-                // console.log(res.success);
-                // if (res.success == true) {
-                //     $('#ajaxImgUpload').attr('src', 'https://via.placeholder.com/300');
-                //     $('#alertMsg').html(res.msg);
-                //     $('#alertMessage').show();
-                // } else if (res.success == false) {
-                //     $('#alertMsg').html(res.msg);
-                //     $('#alertMessage').show();
-                // }
-                // setTimeout(function () {
-                //     $('#alertMsg').html('');
-                //     $('#alertMessage').hide();
-                // }, 4000);
-                // $('.uploadBtn').html('Upload');
-                // $('.uploadBtn').prop('Enabled');
-                // document.getElementById("upload_image_form").reset();
+                if (!response.success) {
+                    openPopupErrors('popup-error', response, status, request);
+                    return;
+                }
+                let data = response.data;
+                image_file_ids.push(data.id);
+
+                $('.slider-wrap .slick').slick('slickAdd', `
+                <div class="slick-element"
+                     style="background: url('/image-file/${data.id}') no-repeat center; background-size: cover; font-size: 0;">
+                    Slider #${data.id}
+                    <div class="slick-element-hover">
+                        <a href="javascript:deleteImage(${data.id})"
+                           class="button delete-image black">
+                            <img src="/asset/images/icon/cancel_white.png"/>
+                        </a>
+                    </div>
+                </div>`);
+                // reset input file
+                input.type = ''
+                input.type = 'file'
             },
             error: function (response, status, error) {
+                openPopupErrors('popup-error', response, status, error);
+                // reset input file
+                input.type = ''
+                input.type = 'file'
             },
         });
     }
+
+    window.onbeforeunload = function () {
+        $.ajax({
+            type: 'POST',
+            url: `/api/image-file/refresh/<?= $identifier ?>`,
+            dataType: 'json',
+        });
+    };
 </script>
